Ensure server does not respond with 500 when action is unknown
